Moved instance restoration to its own hook and into ProtectedRoute
