+++ src/components/ProtectedRoute.tsx
@@ -2,38 +2,31 @@
 //
 // SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
 
-import { ReactNode, useEffect, useState } from 'react';
+import { ReactNode } from 'react';
 import { Navigate, useLocation } from 'react-router-dom';
 import ROUTES from '../definitions/Routes';
-import { useInstance } from '../hooks/useAPI';
 import { RouteRequirements } from '../definitions/RouteObjectList';
 import LoadingSpinner from './LoadingSpinner';
 import 'styles/components/ProtectedRoute.scss';
+import { useRestoreInstance } from '../hooks/useRestoreInstance';
 
 interface ProtectedRouteProps {
     children: ReactNode;
 }
 
 const ProtectedRoute = ({ children }: ProtectedRouteProps) => {
-    const { data: instance, isLoading } = useInstance();
+    const { instance, isLoading, hasCreatedInstance } = useRestoreInstance();
     const location = useLocation();
-    const [hasCreatedInstance, setHasCreatedInstance] = useState<boolean>(false);
 
     const currentRoute = RouteRequirements[location.pathname];
     const needsProfiler = currentRoute?.needsProfilerReport ?? false;
     const needsPerformance = currentRoute?.needsPerformanceReport ?? false;
 
-    useEffect(() => {
-        if (instance) {
-            setHasCreatedInstance(true);
-        }
-    }, [instance]);
-
     if (isLoading && !hasCreatedInstance) {
         return (
             <div className='instance-loader'>
                 <LoadingSpinner />
-                <p>Fetching instance...</p>
+                <p>Fetching/restoring instance...</p>
             </div>
         );
     }
